fixing invalid data bugs

+++ WebApplication1/Services/HeroService.cs
@@ -15,18 +15,34 @@ namespace WebApplication1.Services
         }
         public async Task<Herobio> GetHero(int id)
         {
+            //The method will search db, if there was requested data it will return it else it will use incoming Api to get and set data in database
             var res = _heroRepo.GetHeroTbl(id);
             if (res.Result == null)
             {
                 _heroApiService.userid = id;
                 var ApiTbl = await _heroApiService.Get();
-                _heroRepo.CreateHero(ApiTbl);
-                var show = _heroRepo.GetHeroTbl(id);
-                return await show;
+                if(ApiTbl.id == 0)
+                {
+                    var show = _heroRepo.GetHeroTbl(id);
+                    return await show;
+                }
+                else
+                {
+                    _heroRepo.CreateHero(ApiTbl);
+                    var show = _heroRepo.GetHeroTbl(id);
+                    return await show;
+                } 
+                
             }
             return await res;
-
         }
+        public async Task<Herobio> GetHeroDB(int id)
+        {
+            //just for checking db
+            var show = _heroRepo.GetHeroTbl(id);
+            return await show;
+        }
+
         public async Task<List<Herobio>> Get()
         {
             return await _heroRepo.GetHeros();
@@ -36,9 +52,21 @@ namespace WebApplication1.Services
             _heroRepo.CreateHero(hero);
         }
 
-        public async Task Delete(int id)
+        public  Task Delete(int id)
         {
-            _heroRepo.DeleteHero(id);
+            //This method will check DB at first .if there is a data it will wipe it.
+            var res = _heroRepo.GetHeroTbl(id);
+            if (res.Result == null)
+            {
+                return null;
+            }
+            else
+            {
+                _heroRepo.DeleteHero(id);
+                return _heroRepo.GetHeroTbl(id);
+            }
+            
+
         }
 
         public async Task GetHeros()
@@ -47,6 +75,12 @@ namespace WebApplication1.Services
         }
         public async Task<List<Herobio>> Update(Herobio hero)
         {
+            //This method will check DB at first .if there is a data it will update it.
+            var res = _heroRepo.GetHeroTbl(hero.id);
+            if (res.Result == null)
+            {
+                return null;
+            }
             return await _heroRepo.Update(hero);
         }
     }
